requirements.txt
appdirs
astor
audioread
backports.weakref
bleach
certifi
cffi
chardet
click
colorama
contextlib2
cycler
dataclasses
decorator
dominate
easydict
face-recognition
face-recognition-models
ffmpeg-python
filelock
future
fvcore
h5py
html5lib
hyperopt
idna
imageio
imageio-ffmpeg
importlib-metadata
iopath
joblib
json-tricks
Keras
kiwisolver
librosa
llvmlite
lpips
Markdown
matplotlib
munch
natsort
networkx
nni
numba
numpy
olefile
opencv-python
packaging
Pillow
pooch
portalocker
prettytable
progress
protobuf
psutil
pycparser
pymongo
pyparsing
python-dateutil
PythonWebHDFS
PyWavelets
PyYAML
requests
resampy
responses
ruamel.yaml
ruamel.yaml.clib
schema
scikit-image
scikit-learn
scipy
simplejson
six
SoundFile
tabulate
termcolor
thop
threadpoolctl
tifffile
torch-fidelity
torchsummary
tqdm
urllib3
wcwidth
websockets
Werkzeug
yacs
zipp